[#41] Fix comments on viewstate var naming
