+++ shared/src/commonMain/kotlin/co/nimblehq/blisskmmic/presentation/modules/surveydetail/SurveyQuestionViewState.kt
@@ -1,0 +1,10 @@
+package co.nimblehq.blisskmmic.presentation.modules.surveydetail
+
+data class SurveyQuestionViewState(
+    val isShowingSubmit: Boolean = false,
+    val isLoading: Boolean = false,
+    val currentQuestionIndex: Int = 0,
+    val isShowingSuccess: Boolean = false
+) {
+    constructor() : this(currentQuestionIndex = 0)
+}
